requirements.txt
crewai
crewai-tools
openai
faiss-cpu
langchain
PyPDF2
tqdm
python-dotenv